+++ src/discoverServices.js
@@ -1,5 +1,5 @@
 import { some } from 'lodash';
-import { config } from 'stripes-config';
+import { useStripes } from './StripesContext';
 
 function getHeaders(tenant, token) {
   return {
@@ -227,8 +227,9 @@ function fetchModules(store) {
  * (e.g. not attempting to fetch loan information for a
  * non-circulating library that doesn't provide the circ interface)
  */
-export function discoverServices(store) {
+export function discoverServices(store, config) {
   const promises = [];
+
   if (config.tenantOptions) {
     promises.push(fetchApplicationDetails(store));
     promises.push(fetchGatewayVersion(store));