+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 * Define a permission-set containing values necessary for session-init API requests. Refs STCOR-1031.
 * FOLIO Landing Page > Display text to link to FOLIO release notes. Refs STCOR-1017.
 * Settings > System Information > Add a link to Release Notes. Refs STCOR-1018.
-* Read Stripes Hub stripes-config data from localforage if available. Refs STCOR-1039.
+* Read Stripes Hub stripes-config data from localStorage if available. Refs STCOR-1039.
 
 ## [11.0.0](https://github.com/folio-org/stripes-core/tree/v11.0.0) (2025-02-24)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v10.2.0...v11.0.0)